TESTING IMPROVEMENTS
Improved junit test environment by creating a stock testing database.

This is located at data/Test.db

This is a limited db which is only used to to create relationships with
the junit parameterized tests.

The test will copy the database, and use the copy, deleting it at
cleanup.

Fixed a variety of bugs. Must substantially handling Date and Boolean
objects by the Table_Access class.

I also corrected an error in the Loans table schema. I had to export new
versions of the working databases because of this. Those are included
here.

With this commit, all of the data structures have been tested for basic
insert and read operations. I haven't stress tested them; there may be
edge cases and such that still break them, but the basic functionality
is all good.